Image copy added! Also split out _plot1d_snap.py
