+++ docs/architecture.md
@@ -65,8 +65,11 @@ dialogs, and base context-menu actions. Plot-type-specific snapping and stats
 live in `plot1d.py` and `plot2d.py`.
 
 `src/qplot/windows/plot1d.py` extends the shared plot window for line plots. It
-owns main line rendering, snap-to-trace behavior, and line-plot marquee
-statistics.
+owns main line rendering and line-plot marquee statistics.
+
+`src/qplot/windows/_plot1d_snap.py` contains the line-plot snap-to-trace mixin.
+It owns the snap shortcut/menu action, nearest-point lookup, snap status
+readout, and snap marker display.
 
 `src/qplot/windows/_plot1d_traces.py` contains the line-plot trace mixin. It
 owns secondary trace controls, added-trace refresh handling, right-axis viewbox
